requirements.txt
python-binance>=1.0.19
websockets>=12.0
python-telegram-bot>=20.7
rich>=13.7.0
python-dotenv>=1.0.0
naiohttp>=3.9.0
numpy>=1.24.0
asyncio>=3.4.3
colorama>=0.4.6
pytz>=2023.3
schedule>=1.2.0
asyncpg>=0.29.0
Flask>=2.3.0